pyproject.toml
[project]
name = "djangochatapp"
version = "0.1.0"
description = "A Django and react Chat Project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.6",
    "django-browser-reload>=1.18.0",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-browser-reload>=1.18.0",
    "flake8>=7.1.2",
]
